requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0

# ML/AI
torch>=2.0
transformers>=4.35
datasets>=2.14
scikit-learn>=1.3

# API
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.5

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.21
httpx>=0.25
ruff>=0.1
pre-commit>=3.0

# Benchmarking
matplotlib>=3.8
pandas>=2.1
tabulate>=0.9
psutil>=5.9
